Start camio: dtrace script written and very very basic filtering
Create camio.lua, which parses the command line and modifies the cam.d
script to do what that command line said to do. Currently, there's three
filters: scsi, ata and nvme. These print all the scsi, ata and nvme
commmands respective. The filter is responsible for morphing the generic
cam.d script into something that implements the filter's desired
outcomes. In the future additional arguments will reduce the number of
items that are shown.

We then fork dtrace with the newly created script (using sudo). It will
print a lot of ugly output since we don't (yet) filter that. The next
set of commits will move from using os.execute to io.popen and then read
things in. The plan is to pre-filter out the junk and pass the remainder
onto a post-filter function for the selected filter.

We currently trap the submission path. We likely should trap the
completion path so we can also report results.

Sponsored by:		Netflix