camio: Port over parser.c version of extract
Make this work for multiple words. We do a bit of a hack. NVME is decoed
by word and is little endian. However SCSI and ATA are decoded by byte
and are big endian. Make the decoder decode using LE convention for
word-based and BE using byte-based.

Sponsored by:		Netflix